+++ kfinance/meta_classes.py
@@ -7,7 +7,7 @@ from cachetools import LRUCache, cached
 import numpy as np
 import pandas as pd
 
-from .constants import LINE_ITEMS, BusinessRelationshipType, PeriodType
+from .constants import LINE_ITEMS, BusinessRelationshipType, PeriodType, SegmentType
 from .fetch import KFinanceApiClient
 
 
@@ -303,6 +303,100 @@ class CompanyFunctionsMetaClass:
             )["market_caps"]
         )
         return df.set_index("date")[[column_to_extract]].apply(pd.to_numeric).replace(np.nan, None)
+
+    def _segments(
+        self,
+        segment_type: SegmentType,
+        period_type: Optional[PeriodType] = None,
+        start_year: Optional[int] = None,
+        end_year: Optional[int] = None,
+        start_quarter: Optional[int] = None,
+        end_quarter: Optional[int] = None,
+    ) -> dict:
+        """Get the company's segments"""
+        try:
+            self.validate_inputs(
+                start_year=start_year,
+                end_year=end_year,
+                start_quarter=start_quarter,
+                end_quarter=end_quarter,
+            )
+        except ValueError:
+            return {}
+
+        return self.kfinance_api_client.fetch_segments(
+            company_id=self.company_id,
+            segment_type=segment_type,
+            period_type=period_type,
+            start_year=start_year,
+            end_year=end_year,
+            start_quarter=start_quarter,
+            end_quarter=end_quarter,
+        )["segments"]
+
+    def business_segments(
+        self,
+        period_type: Optional[PeriodType] = None,
+        start_year: Optional[int] = None,
+        end_year: Optional[int] = None,
+        start_quarter: Optional[int] = None,
+        end_quarter: Optional[int] = None,
+    ) -> dict:
+        """Retrieves the templated line of business segments for a given period_type, start_year, start_quarter, end_year and end_quarter.
+
+        :param period_type: The period_type requested for. Can be “annual”, “quarterly”, "ytd". Defaults to “annual” when start_quarter and end_quarter are None.
+        :type start_year: PeriodType, optional
+        :param start_year: The starting calendar year, defaults to None
+        :type start_year: int, optional
+        :param end_year: The ending calendar year, defaults to None
+        :type end_year: int, optional
+        :param start_quarter: The starting calendar quarter, defaults to None
+        :type start_quarter: int, optional
+        :param end_quarter: The ending calendar quarter, defaults to None
+        :type end_quarter: int, optional
+        :return: A dictionary containing the templated line of business segments for each time period, segment name, line item, and value.
+        :rtype: dict
+        """
+        return self._segments(
+            segment_type=SegmentType.business,
+            period_type=period_type,
+            start_year=start_year,
+            end_year=end_year,
+            start_quarter=start_quarter,
+            end_quarter=end_quarter,
+        )
+
+    def geographic_segments(
+        self,
+        period_type: Optional[PeriodType] = None,
+        start_year: Optional[int] = None,
+        end_year: Optional[int] = None,
+        start_quarter: Optional[int] = None,
+        end_quarter: Optional[int] = None,
+    ) -> dict:
+        """Retrieves the templated geographic segments for a given period_type, start_year, start_quarter, end_year and end_quarter.
+
+        :param period_type: The period_type requested for. Can be “annual”, “quarterly”, "ytd". Defaults to “annual” when start_quarter and end_quarter are None.
+        :type start_year: PeriodType, optional
+        :param start_year: The starting calendar year, defaults to None
+        :type start_year: int, optional
+        :param end_year: The ending calendar year, defaults to None
+        :type end_year: int, optional
+        :param start_quarter: The starting calendar quarter, defaults to None
+        :type start_quarter: int, optional
+        :param end_quarter: The ending calendar quarter, defaults to None
+        :type end_quarter: int, optional
+        :return: A dictionary containing the templated geographic segments for each time period, segment name, line item, and value.
+        :rtype: dict
+        """
+        return self._segments(
+            segment_type=SegmentType.geographic,
+            period_type=period_type,
+            start_year=start_year,
+            end_year=end_year,
+            start_quarter=start_quarter,
+            end_quarter=end_quarter,
+        )
 
 
 for line_item in LINE_ITEMS: